start components & navbar
