+++ package.json
@@ -41,6 +41,7 @@
     "axios": "^0.19.0",
     "babel-loader": "^8.0.6",
     "dotenv": "^8.1.0",
+    "dotenv-webpack": "^1.7.0",
     "express": "^4.17.1",
     "google-map-react": "^1.1.4",
     "mdbreact": "^4.19.2",
