+++ src/components/Home.js
@@ -1,0 +1,12 @@
+import React from "react";
+import ListPlaces from "./ListPlaces";
+import NavBar from "./NavBar";
+
+export default function Home({ match }) {
+  return (
+    <div>
+      <NavBar />
+      <ListPlaces match={match} />
+    </div>
+  );
+}
